Allow SstFileReader open file with global seqno
Summary:
Since `SstFileReader` don't know largest seqno of a file, it will fail
this check when it open a file with global seqno:
https://github.com/facebook/rocksdb/blob/ca89ac2ba997dfa0e135bd75d4ccf6f5774a7eff/table/block_based_table_reader.cc#L730
Changes:
* Pass largest_seqno=kMaxSequenceNumber from `SstFileReader` and allow
it to bypass the above check.
* `BlockBasedTable::VerifyChecksum` also double check if checksum will
match when excluding global seqno (this is to make the new test in
sst_table_reader_test pass).

Test Plan:
Added a new test.